Update DetailData.jsx
Added fetch for meteor strike data and and useEffect to try to limit the number of times fetch is called to one. However, it still seems to get called twice? I think the fetch needs to be moved to the App.jsx component anyway because all the child components need it