Update DetailData.jsx
API calls were returning status 429 codes (too many calls). To debug, I added comments to the useEffect where the API calls are made. I removed locationMemo as a dependency on the useEffect. I think this was causing redundant calls to the API, since the locationMemo is updated in the try block where the API calls are made